requirements.txt
pyTelegramBotAPI
yt-dlp
edge-tts
python-dotenv